:bug: solve the undefined behavior when the offset is negative by using signed integers instead of size_t

+++ src/core/pointers.hpp
@@ -96,7 +96,7 @@ class m_ptr {
     T* operator->() const { return data_; }
 
     /** @brief operator *, return the associated object */
-    T& operator[](const size_t idx) const { return data_[idx]; }
+    T& operator[](const int idx) const { return data_[idx]; }
 };
 
 /** @brief data pointer type = root of the data, i.e. the adress of (0,0,0)
